Add wide hou+ menu option for translators
On hou, in the UI file, there are two images for the menu buttons.
One is very wide (button hovered), and one only is as long as the english text
(button not hovered). Translators may be unable to fit the text in the short
button, so I've added a mode which only uses the long button, and uses the engine
to give a fade effect when you hover the button.

Then translators can use the wide button to fit in more text.